Implemented total cpu usage

+++ include/cpu.h
@@ -4,6 +4,6 @@
 #include <cstdint>
 
 uint8_t get_core_count();
-uint8_t get_cpu_usage();
+double get_cpu_usage();
 
 #endif // CPU_H_
